Constraints

PowerModelsProtection.constraint_bus_fault_current — Method

Calculates the fault current at a bus

PowerModelsProtection.constraint_bus_fault_current — Method

constraint for the fault current at the fault_bus

PowerModelsProtection.constraint_current_balance — Method

Calculates the current balance at the non-faulted buses

PowerModelsProtection.constraint_current_balance — Method

Constraint to calculate the fault current at a bus and the current at other buses

PowerModelsProtection.constraint_fault_current_balance — Method

Calculates the current balance at the faulted bus

PowerModelsProtection.constraint_gen_voltage_drop — Method

States that the bus voltage is equal to the internal voltage minus voltage drop across subtransient impedance

PowerModelsProtection.constraint_gen_voltage_drop — Method

Constraint that sets the terminal voltage basd on the internal voltage and the stator impedence

PowerModelsProtection.constraint_i_inverter — Method

Constraints for fault current contribution of inverter in grid-following mode assuming that the inverter current regulating loop operates slowly

PowerModelsProtection.constraint_i_inverter_vs — Method

Constraints for fault current contribution of inverter in grid-following mode assuming that the inverter current regulating loop operates slowly

PowerModelsProtection.constraint_mc_bus_fault_current — Method

PowerModelsProtection.constraint_mc_bus_fault_current — Method

Calculates the current at the faulted bus for multiconductor

PowerModelsProtection.constraint_mc_current_balance — Method

Constraint to calculate the fault current at a bus and the current at other buses for multiconductor

PowerModelsProtection.constraint_mc_fault_current_balance — Method

Calculates the current balance at the faulted bus for multiconductor

PowerModelsProtection.constraint_mc_gen_power_setpoint_imag — Method

generator reactive power setpoint constraint

PowerModelsProtection.constraint_mc_gen_power_setpoint_imag — Method

generator reactive power setpoint constraint

PowerModelsProtection.constraint_mc_gen_voltage_drop — Method

Constraint that sets the terminal voltage basd on the internal voltage and the stator impedence

PowerModelsProtection.constraint_mc_gen_voltage_drop — Method

Constraint that sets the terminal voltage basd on the internal voltage and the stator impedence for multiconductor

PowerModelsProtection.constraint_mc_grid_formimg_inverter_impedance — Method

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_grid_formimg_inverter_virtual_impedance — Method

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_grid_forming_inverter — Method

Constraints for fault current contribution of multiconductor inverter in grid-forming mode

PowerModelsProtection.constraint_mc_grid_forming_inverter — Method

Constraints for fault current contribution of multiconductor inverter in grid-forming mode

PowerModelsProtection.constraint_mc_grid_forming_inverter_impedance — Method

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_grid_forming_inverter_virtual_impedance — Method

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_i_inverter — Method

Constraints for fault current inverter with current set point

PowerModelsProtection.constraint_mc_pq_inverter — Method

Constraints for fault current contribution of multiconductor inverter in grid-following mode

PowerModelsProtection.constraint_mc_pq_inverter — Method

Constraints for fault current contribution of multiconductor inverter in grid-following mode

PowerModelsProtection.constraint_pf_inverter_vs — Method

Constraints for fault current contribution of inverter in grid-following mode with a real voltage drop to handle low-zero terminal voltages

PowerModelsProtection.constraint_pq_inverter — Method

Constraints for fault current contribution of inverter in grid-following mode with pq set points

PowerModelsProtection.constraint_pq_inverter — Method

Constraints for fault current contribution of inverter in grid-following mode with pq set points

PowerModelsProtection.constraint_pq_inverter_mccormick — Method

McCormick relaxation of constraints for fault current contribution of inverter in grid-following mode

PowerModelsProtection.constraint_pq_inverter_region — Method

Constraints for fault current contribution of inverter in grid-following mode operating at arbitrary power factor. Requires objective term

PowerModelsProtection.constraint_unity_pf_inverter — Method

Constraints for fault current contribution of inverter in grid-following mode operating at unity power factor

PowerModelsProtection.constraint_unity_pf_inverter_disjunctive — Method

Constraints for fault current contribution of inverter in grid-following mode with pseudo-binary for current-limiting

PowerModelsProtection.constraint_unity_pf_inverter_rs — Method

Constraints for fault current contribution of inverter in grid-following mode operating at unity power factor with a series resistance to handle low-zero terminal voltages

PowerModelsProtection.constraint_v_inverter — Method

Constraints for fault current contribution of inverter in grid-forming mode

PowerModelsProtection.constraint_v_inverter — Method

Constraints for fault current contribution of inverter in grid-forming mode