Problems

Solvers

PowerModelsProtection.solve_fault_study — Method

solve_fault_study(case::Dict{String,Any}, solver; kwargs...)

Solves a fault study using all active faults under case["fault"] for transmission (matpower) data sets given an optimization solver

kwargs can be any valid keyword argument for PowerModels' run_model function.

PowerModelsProtection.solve_fault_study — Method

solve_fault_study(case::Dict{String,<:Any}, fault_studies::Dict{String,<:Any}, solver; kwargs...)::Dict{String,Any}

Solves a series fault studies in fault_studies, e.g., generated by build_fault_study for transmission (matpower) data sets given an optimization solver

kwargs can be any valid keyword argument for PowerModels' run_model function.

PowerModelsProtection.solve_fault_study — Method

solve_fault_study(file::String, solver; kwargs...)

Given a file, parses file and solves a series of fault studies

PowerModelsProtection.solve_mc_dg_pf — Method

solve_mc_dg_pf(data::Dict{String,<:Any}, solver; kwargs...)

Run Power Flow Problem with DG

PowerModelsProtection.solve_mc_dg_pf — Method

solve_mc_dg_pf(file::String, solver; kwargs...)

Run Power Flow Problem with DG

PowerModelsProtection.solve_mc_fault_study — Method

solve_mc_fault_study(case::Dict{String,<:Any}, solver; kwargs...)

Function to solve a multiconductor (distribution) fault study given a data set case and optimization solver

kwargs can be any valid keyword argument for PowerModelsDistribution's solve_mc_model

PowerModelsProtection.solve_mc_fault_study — Method

solve_mc_fault_study(case::Dict{String,<:Any}, fault_studies::Dict{String,<:Any}, solver; kwargs...)

Solves a series of fault studies given by fault_studies, e.g., built from build_mc_fault_study.

PowerModelsProtection.solve_mc_fault_study — Method

solve_mc_fault_study(file::String, solver; kwargs...)

Given a file, parses the file, and runs the fault study.

PowerModelsProtection.solve_mc_pf — Method

solve_mc_pf(data::Dict{String,<:Any}, solver; kwargs...)

Run Power Flow Problem with Solar

PowerModelsProtection.solve_mc_pf — Method

solve_mc_pf(file::String, solver; kwargs...)

Run Power Flow Problem with Solar

Builders

PowerModelsProtection.build_fault_study — Method

build_fault_study(data::Dict; default_fault_resistance::Real=0.0001)

Builds a dictionary of fault studies on a transmission (single-phase positive sequence) network that are intended to be used in conjunction with solve_fault_study.

The function will iterate over all buses and create faults using the default fault resistance. The fault study dictionary will have the following structure:

    Dict{String,Any}(
        "bus_i" => Dict{String,Any}(
            "fault_bus" => bus_i
            "gf" => 1 / resistance,
            "status" => 1
        ),
        ...
    )

PowerModelsProtection.build_fault_study — Method

build_fault_study(pm::_PM.AbstractPowerModel)

Builds a transmission fault study optimization problem

PowerModelsProtection.build_mc_dg_pf — Method

build_mc_dg_pf(pm::_PMD.AbstractUnbalancedPowerModel)

Constructor for Power Flow Problem with DG

PowerModelsProtection.build_mc_fault_study — Method

build_mc_fault_study(data::Dict{String,<:Any}; resistance::Real=0.01, phase_resistance::Real=0.01)::Dict{String,Any}

Add all fault type data to model for study for multiconductor networks

PowerModelsProtection.build_mc_fault_study — Method

build_mc_fault_study(pm::_PMD.AbstractUnbalancedPowerModel)

Builds a multiconductor (distribution) fault study optimization problem

PowerModelsProtection.build_mc_pf — Method

build_mc_pf(pm::_PMD.AbstractUnbalancedPowerModel)

Constructor for Power Flow Problem with Solar

PowerModelsProtection.build_mc_sparse_fault_study — Method

build_mc_sparse_fault_study(data::Dict{String,<:Any}; resistance::Real=0.01, phase_resistance::Real=0.01)::Dict{String,Any}

Builds sparse collection of fault studies using a network graph

Solution Helpers

DEPRECIATED Solver functions