Variables

PowerModelsProtection.variable_branch_current — Method

variable_branch_current(pm::_PM.AbstractIVRModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true, kwargs...)

Copies from PowerModels and PowerModelsDistribution without power vars

PowerModelsProtection.variable_bus_fault_current — Method

variable_bus_fault_current(pm::_PM.AbstractIVRModel; nw::Int=nw_id_default, report::Bool=true)

fault current variables for active faults

PowerModelsProtection.variable_gen — Method

variable_gen(pm::_PM.AbstractIVRModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true, kwargs...)

builds generator variables for transmission networks

PowerModelsProtection.variable_gen_loading — Method

variable_gen_loading(pm::_PM.AbstractIVRModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true)

variable: pg[j] for j in gen

PowerModelsProtection.variable_mc_bus_fault_current — Method

variable_mc_bus_fault_current(pm::_PMD.AbstractUnbalancedIVRModel; nw::Int=nw_id_default, report::Bool=true)

variables for multiconductor fault currents for active faults

PowerModelsProtection.variable_mc_grid_formimg_inverter — Method

variable_mc_grid_formimg_inverter(pm::_PMD.AbstractUnbalancedIVRModel; nw::Int=nw_id_default, bounded::Bool=true, kwargs...)

variables for multiconductor grid forming inverters

PowerModelsProtection.variable_mc_pq_inverter — Method

variable_mc_pq_inverter(pm::_PMD.AbstractUnbalancedIVRModel; nw::Int=nw_id_default, bounded::Bool=true, kwargs...)

variables for multiconductor pq inverters

PowerModelsProtection.variable_mc_storage_current — Method

variable_mc_storage_current(pm::_PMD.AbstractUnbalancedIVRModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true)

variables for output terminal currents for grid-connected energy storage

PowerModelsProtection.variable_mc_storage_current_imaginary — Method

variable_mc_storage_current_imaginary(pm::_PMD.AbstractUnbalancedIVRModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true)

variables for real portion of output terminal currents for grid-connected energy storage

PowerModelsProtection.variable_mc_storage_current_real — Method

variable_mc_storage_current_real(pm::_PMD.AbstractUnbalancedIVRModel; nw::Int=nw_id_default, bounded::Bool=true, report::Bool=true)

variables for real portion of output terminal currents for grid-connected energy storage

PowerModelsProtection.variable_mc_storage_grid_forming_inverter — Method

variable_mc_storage_grid_forming_inverter(pm::_PMD.AbstractUnbalancedIVRModel; nw::Int=nw_id_default, bounded::Bool=true, kwargs...)

variables associated with grid-connected energy storage: internal voltage, virutal impedance, dc-link power, faulted state

PowerModelsProtection.variable_pq_inverter — Method

variable_pq_inverter(pm::_PM.AbstractIVRModel; nw::Int=nw_id_default, bounded::Bool=true, kwargs...)

variables for pq inverters