Constraints

PowerModelsProtection.constraint_bus_fault_current — Method

constraint_bus_fault_current(pm::_PM.AbstractIVRModel, nw::Int, i::Int, fault_bus::Int, g::Real)

Calculates the fault current at a bus

PowerModelsProtection.constraint_bus_fault_current — Method

constraint_bus_fault_current(pm::_PM.AbstractIVRModel, i::Int; nw::Int=nw_id_default)

constraint for the fault current at the fault_bus

PowerModelsProtection.constraint_current_balance — Method

constraint_current_balance(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_arcs, bus_gens, bus_gs, bus_bs)

Calculates the current balance at the non-faulted buses

PowerModelsProtection.constraint_current_balance — Method

constraint_current_balance(pm::_PM.AbstractPowerModel, i::Int; nw::Int=nw_id_default)

Constraint to calculate the fault current at a bus and the current at other buses

PowerModelsProtection.constraint_fault_current_balance — Method

constraint_fault_current_balance(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_arcs, bus_gens, bus_gs, bus_bs)

Calculates the current balance at the faulted bus

PowerModelsProtection.constraint_gen_voltage_drop — Method

constraint_gen_voltage_drop(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_id, r, x, vgr, vgi)

States that the bus voltage is equal to the internal voltage minus voltage drop across subtransient impedance

PowerModelsProtection.constraint_gen_voltage_drop — Method

constraint_gen_voltage_drop(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default)

Constraint that sets the terminal voltage basd on the internal voltage and the stator impedence

PowerModelsProtection.constraint_i_inverter — Method

constraint_i_inverter(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default)

Constraints for fault current contribution of inverter in grid-following mode assuming that the inverter current regulating loop operates slowly

PowerModelsProtection.constraint_i_inverter_vs — Method

constraint_i_inverter_vs(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_id::Int, vs, pg, qg, cm)

Constraints for fault current contribution of inverter in grid-following mode assuming that the inverter current regulating loop operates slowly

PowerModelsProtection.constraint_mc_bus_fault_current — Method

constraint_mc_bus_fault_current(pm::_PMD.AbstractUnbalancedIVRModel, i::Int; nw::Int=nw_id_default)

Constraint for Kirchoff's current law on faulted buses

PowerModelsProtection.constraint_mc_bus_fault_current — Method

constraint_mc_bus_fault_current(pm::_PMD.AbstractUnbalancedPowerModel, nw::Int, i::Int, bus::Int, connections::Vector{Int}, Gf::Matrix{<:Real}, Bf::Matrix{<:Real})

Calculates the current at the faulted bus for multiconductor

PowerModelsProtection.constraint_mc_current_balance — Method

constraint_mc_current_balance(pm::_PMD.AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default)

Constraint to calculate the fault current at a bus and the current at other buses for multiconductor

PowerModelsProtection.constraint_mc_fault_current_balance — Method

constraint_mc_fault_current_balance(pm::_PMD.AbstractUnbalancedIVRModel, nw::Int, i::Int, fault::Int, terminals::Vector{Int}, grounded::Vector{Bool}, bus_arcs::Vector{Tuple{Tuple{Int,Int,Int},Vector{Int}}}, bus_arcs_sw::Vector{Tuple{Tuple{Int,Int,Int},Vector{Int}}}, bus_arcs_trans::Vector{Tuple{Tuple{Int,Int,Int},Vector{Int}}}, bus_gens::Vector{Tuple{Int,Vector{Int}}}, bus_storage::Vector{Tuple{Int,Vector{Int}}}, bus_shunts::Vector{Tuple{Int,Vector{Int}}})

Calculates the current balance at the faulted bus for multiconductor

PowerModelsProtection.constraint_mc_gen_power_setpoint_imag — Method

constraint_mc_gen_power_setpoint_imag(pm::_PMD.AbstractUnbalancedPowerModel, n::Int, i, qg)

generator reactive power setpoint constraint

PowerModelsProtection.constraint_mc_gen_power_setpoint_imag — Method

constraint_mc_gen_power_setpoint_imag(pm::_PMD.AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default, kwargs...)

generator reactive power setpoint constraint

PowerModelsProtection.constraint_mc_gen_voltage_drop — Method

constraint_mc_gen_voltage_drop(pm::_PMD.AbstractUnbalancedIVRModel, n::Int, i::Int, bus_id::Int, r, x, vgr, vgi, terminals)

Constraint that sets the terminal voltage basd on the internal voltage and the stator impedence

PowerModelsProtection.constraint_mc_gen_voltage_drop — Method

constraint_mc_gen_voltage_drop(pm::_PMD.AbstractUnbalancedPowerModel; nw::Int=nw_id_default)

Constraint that sets the terminal voltage basd on the internal voltage and the stator impedence for multiconductor

PowerModelsProtection.constraint_mc_grid_formimg_inverter_impedance — Method

constraint_mc_grid_formimg_inverter_impedance(pm::_PMD.AbstractUnbalancedIVRModel, nw::Int, i::Int, bus_id::Int, vr0, vi0, r, x, pmax, cmax)

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_grid_formimg_inverter_virtual_impedance — Method

constraint_mc_grid_formimg_inverter_virtual_impedance(pm::_PMD.AbstractUnbalancedIVRModel, nw::Int, i::Int, bus_id::Int, vr0, vi0, pmax, cmax, smax, ang, terminals)

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_grid_forming_inverter — Method

constraint_mc_grid_forming_inverter(pm::_PMD.AbstractUnbalancedIVRModel, nw::Int, i::Int, bus_id::Int, vrstar, vistar, pmax, cmax)

Constraints for fault current contribution of multiconductor inverter in grid-forming mode

PowerModelsProtection.constraint_mc_grid_forming_inverter — Method

constraint_mc_grid_forming_inverter(pm::_PMD.AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default)

Constraints for fault current contribution of multiconductor inverter in grid-forming mode

PowerModelsProtection.constraint_mc_grid_forming_inverter_impedance — Method

constraint_mc_grid_forming_inverter_impedance(pm::_PMD.AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default)

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_grid_forming_inverter_virtual_impedance — Method

constraint_mc_grid_forming_inverter_virtual_impedance(pm::_PMD.AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default)

Constraints for fault current contribution of multiconductor inverter in grid-forming mode with power matching

PowerModelsProtection.constraint_mc_i_inverter — Method

constraint_mc_i_inverter(pm::_PMD.AbstractUnbalancedIVRModel, n::Int, i::Int, bus_id::Int, pg, qg, cmax)

Constraints for fault current inverter with current set point

PowerModelsProtection.constraint_mc_pq_inverter — Method

constraint_mc_pq_inverter(pm::_PMD.AbstractUnbalancedIVRModel, nw::Int, i::Int, bus_id::Int, pg, qg, cmax)

Constraints for fault current contribution of multiconductor inverter in grid-following mode

PowerModelsProtection.constraint_mc_pq_inverter — Method

constraint_mc_pq_inverter(pm::_PMD.AbstractUnbalancedPowerModel, i::Int; nw::Int=nw_id_default)

Constraints for fault current contribution of multiconductor inverter in grid-following mode

PowerModelsProtection.constraint_mc_storage_grid_forming_inverter — Method

constraint_mc_storage_grid_forming_inverter(pm::_PMD.AbstractUnbalancedIVRModel, nw::Int, i, bus_id::Int, connections)

Constrants for grid-forming inverter with storage

PowerModelsProtection.constraint_mc_storage_grid_forming_inverter — Method

constraint_mc_storage_grid_forming_inverter(pm::_PMD.AbstractUnbalancedIVRModel, i::Int; nw::Int=nw_id_default)

Constraint for fault-current contribution battery energy storage inverters

PowerModelsProtection.constraint_pf_inverter_vs — Method

constraint_pf_inverter_vs(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_id::Int, vs, pg, qg, cmax)

Constraints for fault current contribution of inverter in grid-following mode with a real voltage drop to handle low-zero terminal voltages

PowerModelsProtection.constraint_pq_inverter — Method

constraint_pq_inverter(pm::_PM.AbstractIVRModel, nw::Int, i::Int, bus_id::Int, pg, qg, cmax)

Constraints for fault current contribution of inverter in grid-following mode with pq set points

PowerModelsProtection.constraint_pq_inverter — Method

constraint_pq_inverter(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default)

Constraints for fault current contribution of inverter in grid-following mode with pq set points

PowerModelsProtection.constraint_pq_inverter_mccormick — Method

constraint_pq_inverter_mccormick(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_id::Int, pg, qg, cmax)

McCormick relaxation of constraints for fault current contribution of inverter in grid-following mode

PowerModelsProtection.constraint_pq_inverter_region — Method

constraint_pq_inverter_region(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_id::Int, pg, qg, cmax)

Constraints for fault current contribution of inverter in grid-following mode operating at arbitrary power factor. Requires objective term

PowerModelsProtection.constraint_unity_pf_inverter — Method

constraint_unity_pf_inverter(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_id::Int, pg, qg, cmax)

Constraints for fault current contribution of inverter in grid-following mode operating at unity power factor

PowerModelsProtection.constraint_unity_pf_inverter_disjunctive — Method

constraint_unity_pf_inverter_disjunctive(pm::_PM.AbstractIVRModel, nw::Int, i::Int, bus_id::Int, pg, qg, cmax)

Constraints for fault current contribution of inverter in grid-following mode with pseudo-binary for current-limiting

PowerModelsProtection.constraint_unity_pf_inverter_rs — Method

constraint_unity_pf_inverter_rs(pm::_PM.AbstractIVRModel, n::Int, i::Int, bus_id::Int, r, pg, qg, cm)

Constraints for fault current contribution of inverter in grid-following mode operating at unity power factor with a series resistance to handle low-zero terminal voltages

PowerModelsProtection.constraint_v_inverter — Method

constraint_v_inverter(pm::_PM.AbstractIVRModel, n::Int, i, bus_id, r, x, vgr, vgi, cmax)

Constraints for fault current contribution of inverter in grid-forming mode

PowerModelsProtection.constraint_v_inverter — Method

constraint_v_inverter(pm::_PM.AbstractPowerModel; nw::Int=nw_id_default)

Constraints for fault current contribution of inverter in grid-forming mode