Internal Functions

PowerModelsProtection._pmp_dimensionalize_math_extensions — Constant

pointers for fields to get unit transformations back to si units in solutions

PowerModelsProtection._pmp_eng2math_passthrough — Constant

field/values to passthrough from the ENGINEERING to MATHEMATICAL data models

PowerModelsProtection._pmp_map_math2eng_extensions — Constant

necessary function reference dict for solution transformations

PowerModelsProtection._add_dict — Method

Function to help add keyword arguments to the circuit information dictionary.

PowerModelsProtection._bisection — Method

Bisection solver used to help fit a curve to provided points on tcc_curves.

PowerModelsProtection._check_keys — Method

Function to apply haskey function to all elements in a vector. Used for checking if all cts are in the circuit when adding a differential relay. Returns a Bool

PowerModelsProtection._create_3p_fault — Method

creates a three-phase fault

PowerModelsProtection._create_3pg_fault — Method

creates a three-phase-ground fault

PowerModelsProtection._create_lg_fault — Method

creates a line-ground fault

PowerModelsProtection._create_ll_fault — Method

creates a line-line fault

PowerModelsProtection._create_llg_fault — Method

creates a line-line-ground fault

PowerModelsProtection._differential_time — Method

Function that calculates short-invers operation time of relay. Uses restraint instead of tap setting.

PowerModelsProtection._dss2eng_ct! — Method

Helper function to convert dss data for monitors to engineering current transformer model.

PowerModelsProtection._dss2eng_curve! — Method

Helper function for converting dss tcc_curves to engineering model

PowerModelsProtection._dss2eng_fuse! — Method

Helper function for converting dss fuse to engineering fuse

PowerModelsProtection._dss2eng_gen_dynamics! — Method

helper function to build extra dynamics information for generator or vsource objects

PowerModelsProtection._dss2eng_relay! — Method

Helper function for converting dss relay to engineering relay model.

PowerModelsProtection._dss2eng_solar_dynamics! — Method

helper function to build extra dynamics information for pvsystem objects

PowerModelsProtection._eng2math_fault! — Method

helper function to transformer fault objects from ENGINEERING to MATHEMATICAL data models

PowerModelsProtection._eng2math_protection! — Method

helper function to transform protection equipment objects from ENGINEERING to MATHEMATICAL data models

PowerModelsProtection._fuse_operation — Method

Determines whether a fuse operates given the current through the line it is protecting

PowerModelsProtection._get_current — Method

Function to get current that is flowing through the relay or fuse we are looking at. Gets the type of relay from dictionary then uses the proper method for calculating current through relay operating coil. Outputs a vector.

PowerModelsProtection._get_current_math — Method

Same as getcurrent but uses mathematical model.

PowerModelsProtection._get_phase_base_order — Method

helper function to get faulted phases base order

PowerModelsProtection._interpolate_time — Method

Function to interpolate time values from tcc_curve

PowerModelsProtection._long_time — Method

Function that calculates the long-inverse time of a relay for delayed trips.

PowerModelsProtection._map_math2eng_fault! — Method

helper function to transform fault solutions back into ENGINEERING from MATHEMATICAL data models

PowerModelsProtection._map_math2eng_protection! — Method

Helper function to convert tripped relays and blown fuses from MATHEMATICAL to ENGINEERING models

PowerModelsProtection._non_ideal_ct — Method

_non_ideal_ct(relay_data,CT_data,Iabc)

Converts primary side current to the actual current going through relay coil based on non-ideal parameters. Unused.

PowerModelsProtection._p_to_s — Method

Function to convert phase values to sequence values. Takes vector of 3 phases, returns vector of 3 sequences.

PowerModelsProtection._rebase_pu_fault! — Method

function to rebase units into pu on fault objects

PowerModelsProtection._rebase_pu_gen_dynamics! — Method

function to rebase extra fields for generator dynamics into per unit

PowerModelsProtection._ref_add_fault! — Method

Adds the fault to the model

PowerModelsProtection._ref_add_grid_forming_bus! — Method

identifies grid forming buses

PowerModelsProtection._ref_add_mc_fault! — Method

Adds the fault to the model for multiconductor

PowerModelsProtection._ref_add_mc_solar! — Method

Calculates the power from solar based on inputs

PowerModelsProtection._ref_add_mc_storage! — Method

Add battery energy storage to the model

PowerModelsProtection._relay_operation — Method

Determines whether a relay operates given the current through its protected object

PowerModelsProtection._restraint — Method

Function that calculated the slope setting of differential relay based on tap setting.

PowerModelsProtection._short_time — Method

Function that calculates the short-inverse time of a relay for fast trips.

PowerModelsProtection._solution_fs! — Method

adds additional variable transformations for fault study solutions of distribution networks

PowerModelsProtection._solution_pm_fs! — Method

adds additional variable transformations for fault study solutions of transmission networks

PowerModelsProtection._solution_protection! — Method

calculates relay and fuse operation times and adds them to solution dictionary