Internal Functions

PowerModelsProtection._pmp_dimensionalize_math_extensions — Constant

pointers for fields to get unit transformations back to si units in solutions

PowerModelsProtection._pmp_eng2math_passthrough — Constant

field/values to passthrough from the ENGINEERING to MATHEMATICAL data models

PowerModelsProtection._pmp_map_math2eng_extensions — Constant

necessary function reference dict for solution transformations

PowerModelsProtection._create_3p_fault — Method

creates a three-phase fault

PowerModelsProtection._create_3pg_fault — Method

creates a three-phase-ground fault

PowerModelsProtection._create_lg_fault — Method

creates a line-ground fault

PowerModelsProtection._create_ll_fault — Method

creates a line-line fault

PowerModelsProtection._create_llg_fault — Method

creates a line-line-ground fault

PowerModelsProtection._dss2eng_gen_dynamics! — Method

helper function to build extra dynamics information for generator or vsource objects

PowerModelsProtection._dss2eng_solar_dynamics! — Method

helper function to build extra dynamics information for pvsystem objects

PowerModelsProtection._eng2math_fault! — Method

helper function to transformer fault objects from ENGINEERING to MATHEMATICAL data models

PowerModelsProtection._get_phase_base_order — Method

helper function to get faulted phases base order

PowerModelsProtection._map_math2eng_fault! — Method

helper function to transform fault solutions back into ENGINEERING from MATHEMATICAL data models

PowerModelsProtection._rebase_pu_fault! — Method

function to rebase units into pu on fault objects

PowerModelsProtection._rebase_pu_gen_dynamics! — Method

function to rebase extra fields for generator dynamics into per unit

PowerModelsProtection._ref_add_fault! — Method

Adds the fault to the model

PowerModelsProtection._ref_add_grid_forming_bus! — Method

identifies grid forming buses

PowerModelsProtection._ref_add_mc_fault! — Method

Adds the fault to the model for multiconductor

PowerModelsProtection._ref_add_mc_solar! — Method

Calculates the power from solar based on inputs

PowerModelsProtection._ref_add_mc_storage! — Method

Add battery energy storage to the model

PowerModelsProtection._solution_fs! — Method

adds additional variable transformations for fault study solutions of distribution networks

PowerModelsProtection._solution_pm_fs! — Method

adds additional variable transformations for fault study solutions of transmission networks